Use plugin functionality to document forward models in Everest
