+++ src/everest/plugins/everest_plugin_manager.py
@@ -1,3 +1,5 @@
+from typing import Any, Dict
+
 import pluggy
 
 from everest.plugins import hook_impl, hook_specs
@@ -5,7 +7,7 @@ from everest.strings import EVEREST
 
 
 class EverestPluginManager(pluggy.PluginManager):
-    def __init__(self, plugins=None):
+    def __init__(self, plugins=None) -> None:
         super(EverestPluginManager, self).__init__(EVEREST)
         self.add_hookspecs(hook_specs)
         if plugins is None:
@@ -14,3 +16,7 @@ class EverestPluginManager(pluggy.PluginManager):
         else:
             for plugin in plugins:
                 self.register(plugin)
+
+    def get_documentation(self) -> Dict[str, Any]:
+        docs = self.hook.get_forward_model_documentations()
+        return {k: v for d in docs for k, v in d.items()} if docs else {}
