+++ src/everest/plugins/plugin_response.py
@@ -1,4 +1,8 @@
+from typing import Generic, TypeVar
+
 from decorator import decorator
+
+T = TypeVar("T")
 
 
 @decorator
@@ -11,13 +15,13 @@ def plugin_response(func, plugin_name="", *args, **kwargs):  # pylint: disable=k
     )
 
 
-class PluginResponse:
-    def __init__(self, data, plugin_metadata):
-        self.data = data
-        self.plugin_metadata = plugin_metadata
-
-
 class PluginMetadata:
     def __init__(self, plugin_name, function_name):
         self.plugin_name = plugin_name
         self.function_name = function_name
+
+
+class PluginResponse(Generic[T]):
+    def __init__(self, data: T, plugin_metadata: PluginMetadata) -> None:
+        self.data = data
+        self.plugin_metadata = plugin_metadata
