+++ src/everest/util/__init__.py
@@ -5,7 +5,7 @@ import os
 from ropt.version import version as ropt_version
 
 from ert.shared.version import version as ert_version
-from everest.plugins.hook_manager import EverestPluginManager
+from everest.plugins.everest_plugin_manager import EverestPluginManager
 from everest.strings import DATE_FORMAT, DEFAULT_LOGGING_FORMAT
 from everest.util.async_run import async_run  # noqa
